+++ Cargo.toml
@@ -11,14 +11,13 @@ appveyor = { repository = "pzread/furakus" }
 
 [dependencies]
 bytes = "0.4"
-dotenv = "0.10"
+dotenv = "0.11"
 futures = "0.1"
 hyper = "0.11"
 language-tags = "0.2"
 lazy_static = "1.0"
 mime = "0.3"
 native-tls = "0.1"
-openssl = "0.9"
 regex = "0.2"
 ring = "0.12"
 serde = "1.0"
@@ -28,4 +27,4 @@ tokio-core = "0.1"
 tokio-tls = "0.1"
 unicase = "2.1"
 url = "1.6"
-uuid = { version = "0.5", features = ["v4"] }
+uuid = { version = "0.6", features = ["v4"] }
