+++ src/tls/mod.rs
@@ -1,0 +1,4 @@
+#[cfg(target_os = "windows")]
+mod schannel;
+#[cfg(target_os = "windows")]
+pub use self::schannel::build_tls_from_pfx;
